Universal MPR support
This is an update for the library to enable all 3 types of transfer functions of this series of pressures sensors.

The library now provides ability to zero the sensor at run time.
The library now provides option to update the calibration.
Those settings are defined by the transfer function, however
measuring the calibration will improve performance.

The library now provides option to report RAW sensor counts, which is needed for calibration.

The math to convert sensor counts to pressure has been reworked and will now report negative numbers instead of overflowing to the max scale of uint32_t.

An example program using different sensor types is provided as second example.